+++ pkg/timeDate/NEWS.md
@@ -1,4 +1,29 @@
 
+- the generic `timeDate()` gets argument '...' to allow methods for it to have
+  additional arguments (e.g., for DST gaps).
+
+- the 'character' method for `dateTime()` gets a new argument `dst_gap` to
+  control what to do with non-existent DST times at the requested `FinCenter`
+  with options to add/subtract ("+", "-") the DST shift or set them to `NA`.
+
+- `DateTime()` was not handling correctly some times just after the switch
+  to/from DST. This was visible mostly for time zones away from GMT and GMT+1.
+
+- In `timeSequence()`, if any of the generated times would fall in DST gaps,
+  they are moved by "+1 hour", corresponding to `dst_gap = "+"` in `timeDate`.
+  This is consistent with `seq` for other time objects.  Currently there is no
+  option to change this behaviour of `timeSequence`.
+
+  Previously `timeSequence` was moving DST gaps down by 1 hour (for by =
+  'DSTday' and similar). This was not consistent similar time functions in R and
+  was actually due to a bug (or unfinished DST handling) in `timeDate`, see
+  remarks for `timeDate` above.
+  
+- `timeSequence()` now throws error if argument `from` is in a DST gap. It seems
+  desirable to have a default action for this case. Rolling the faulty time by
+  an hour in the case of 'DSTday' may be suitable in most cases but for other
+  values of `by` it might be totally wrong. 
+
 - London financial centre holidays - fixed and/or included non-standard holidays
   (e.g., Early May Bank holiday was moved in 2020 to VE day; Spring Bank holiday
   was moved in Queen's Jubilee years; state funeral of the Queen).  Millenium